refactor(ui): extract SessionList from App.tsx
Moves the inline session sidebar implementation from App.tsx (five
components plus two helpers, ~320 LOC) into a new tracked
src/components/SessionList.tsx. The extracted surface:

- SessionList (the orientation-aware container — vertical desktop
  sidebar or horizontal mobile strip)
- SessionItem (row with status dot, kind badge, slug, repo chip,
  status chip, preview line)
- KindBadge (P / C / V indicator)
- GroupHeader + GroupView (DAG / parent-child / variant / standalone
  group rendering)
- useFlashOnChange (internal hook — only used by SessionItem)
- statusDot, shortRepo, dagStatusTone (helpers)
- SessionItemKind (public type)

App.tsx now imports SessionList and statusDot from the new module
instead of defining them inline. buildSessionGroups / SessionGroup
are no longer needed at the App.tsx import level since they only
serve the sidebar. Zero behaviour or visual changes — test markup
(session-list, session-row-*, kind-badge-*, group-dag-*) is identical.

Pre-existing but separate: the conductor plan's PR 5 also called for
replacing the emoji-prefix status/attention pills in src/components/
shared.tsx and src/components/ContextMenu.tsx with inline SVG glyphs,
and for overlaying an active-tool chip plus token/cost totals sourced
from the per-session transcript buffer. Those each need design work
(~13 SVGs, plus a lazy-access pattern that avoids fetching every
session's transcript on first render) and are deferred to a follow-up.
MessageInput's palette and shape have already been refreshed by the
PR #33 composer rework, so no additional composer change is needed.
